requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.9
jinja2==3.1.4
pillow==10.4.0
rembg==2.0.59
stripe==11.6.0
